+++ RELEASES.md
@@ -2,6 +2,7 @@
 
 ### Bug fixes
 
+-   Fix GNOME System Monitor icon missing for GNOME 46 and later [[#128](https://github.com/AstraExt/astra-monitor/issues/128)]
 -   Overall Performance improvements
 
 # Astra Monitor 25 - May 8 2024